refactor: main.py CLI 및 유틸리티 기능 분리
main.py의 CLI와 유틸리티 기능을 별도의 디렉터리와 파일로 분리
