+++ cli/__init__.py
@@ -1,0 +1,2 @@
+"""CLI package for argument parsing and terminal output helpers."""
+
