+++ mock_parser.py
@@ -63,19 +63,39 @@ def get_dvwa_mock_surfaces(
         #     cookies=auth_cookies,
         #     description="DVWA File Inclusion (GET)",
         # ),
+        # AttackSurface(
+        #     url=f"{root}/vulnerabilities/brute/",
+        #     method=HttpMethod.GET,
+        #     param_location=ParamLocation.QUERY,
+        #     parameters={
+        #         "username": "admin",
+        #         "password": "password",
+        #         "Login": "Login",
+        #         "user_token": "",
+        #     },
+        #     cookies=auth_cookies,
+        #     description="DVWA Brute Force (GET, with CSRF token)",
+        #     dynamic_tokens=["user_token"],
+        # ),
+        # AttackSurface(
+        #     url=f"{root}/vulnerabilities/brute/",
+        #     method=HttpMethod.GET,
+        #     param_location=ParamLocation.QUERY,
+        #     parameters={
+        #         "username": "admin",
+        #         "password": "password",
+        #         "Login": "Login",
+        #     },
+        #     cookies=auth_cookies,
+        #     description="DVWA Brute Force (GET, no CSRF token)",
+        # ),
         AttackSurface(
-            url=f"{root}/vulnerabilities/brute/",
+            url=f"{root}/vulnerabilities/sqli/",
             method=HttpMethod.GET,
             param_location=ParamLocation.QUERY,
-            parameters={
-                "username": "admin",
-                "password": "password",
-                "Login": "Login",
-                "user_token": "",
-            },
+            parameters={"id": "1", "Submit": "Submit"},
             cookies=auth_cookies,
-            description="DVWA Brute Force (GET)",
-            dynamic_tokens=["user_token"],
+            description="DVWA SQL Injection (GET)",
         ),
         # AttackSurface(
         #     url=f"{root}/vulnerabilities/upload/",
